[BUGFIX] Fix query for getting content in PostgreSQL
In postgres all non-double-quoted column names are converted to lower case before query execution, so 'colPos' becomes 'colpos'. However, because the install tool created the column with double quotes, it is properly named 'colPos' in the database, with big P.

Please test this against other DBMS, I don't have a MySQL instance ready for this.